+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Update README.md
 - Make example more usable
+- Implement node pinning with `controller.setPinned`
+- Implement node replacement with `controller.replaceNode`
 
 ## 0.1.1